preparations for the API
This is in preparation for the API. It should get a handler for the file
directory so that it can push jobs to the file handler.
That way it should also be possible to get the data out of it again and
push aggregation jobs too. But we will see.